#59 - Smoke coming out of cups and refactoring
